% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpModel.R
\name{nlpModel}
\alias{nlpModel}
\title{Create email classifier with Linear Discriminant Analysis}
\usage{
nlpModel(data, labels, verbose = TRUE)
}
\arguments{
\item{data}{A data.frame with all features and labels}

\item{labels}{A vector of matching labels}

\item{verbose}{Print progress messages}
}
\value{
A caret model that predicts spam or ham
}
\description{
Create email classifier with Linear Discriminant Analysis
}
